Keymap Descriptions, Whichkey, Tabs and More
Changes:
- Wanted to install which key, which is a way to get suggestions on what
commands you can technically execute. However, many of my keymaps didnt
have descriptions so I went through them and documented them
- Installed bufferline to allow for tabs like the ones in VSCode.
Experimented a bit with the keymaps for it
- Found out that weirdly enough, tabs are treated more or less the same
as a buffer in bufferline, so I installed scope.vim to give them a bit
of distinction
- Added a fileexplorer separator so that my neotree wouldn't get in the
way of the new buffer tabs
- New plugin for commenting lines! Would love that it would use the same
shortcut I use in VSCode but I couldn't get it working. I don't want to
have so many different shortcuts between the two.
- Installed lazydev, so that my lua config can get a bit of LSP love for
vim methods. Added it as a CMP completion source
- Added lexima for auto closing parenthesis, brackets and others
- Re-configured neotree a bit more so that hidden files would show up,
and a new source selector bar would appear on top of the file tree. I
like how it looks this way.