Fixes and Better Integrations
Changes:
- Modified the command for closing the current buffer by adding a check
to make sure that the neo-tree is open before toggling. If its not open
it will simply close the current buffer
- Removed cmd_action from the LSP config
- Added borders to the mason menu for better readability. Also removed
the unused variable "client" and added a definition for the previously
undefined "bufnr" variable
- Added the extensions for lualine for fugitive, neotree, toggleterm,
mason, lazy, nvim-dap-ui and nvcheatsheet
- Added this option: use_libuv_file_watcher to constantly update the
status of files. Disabled it for now
- Forgot to delete the "other.lua" file. Should be gone now
- Modified the shortcut for listing buffers. Now it should be buffer
list (bl)
- Added borders to the Lazy UI as well