Updating Python LSPs and Configs
Wanted to use neovim for Opus, but it was giving me way too many problems while editing. Ruff was not working and a lot of the code standards were failing, so I had to fix a bunch of things, not only with Neovim, but with Opus itself (but those will be pushed separately).

Changes:
- Removed pylsp. Its not necessary with basedpyright installed as well
- Enabled ruff. It was not being enabled or used, whoops.
- Grouped all of the packages to manage python virtual environments into a single file called "python-venv.lua"
- Deleted the venv-selector plugin. It was not working even though I liked it

Other changes:
- Updated the description in the cheatsheet for the file tree