Trouble and Notify Fixes + Cheatsheet
Writing :Telescope notify was getting a bit tiresome, so I created a "fn" function to display with telescope all of the shown notifications so far. Also reworked the trouble shortcuts to make them a bit simpler and not have them overlap with each other. Documented both changes in the cheatsheet.