Improved Keymaps for Git, Copilot and Comments
Wanted to actually use the gitsigns plugin like I would in VSCode so I
decided to configure it with proper keymaps, but it would conflict with
the keymaps in the comments.nvim plugin, since it was set to
"<leader>gc". Removed all comments.nvim plugins to not use the leader
key, so that it can be done by simply using "gcc". Now there were no
conflicts. Then I changed the keymaps for copilot to be "cco" for
opening the chat and "ccx" for closing it. Its now working as expected.

Documented everything within the cheatsheet