Shorter Comment Blocks
I want my comment blocks to match the width of my existing blocks.
However, in vscode I set the width of the block in its entirety, the
width in the neovim plugin only accounts for the intermediary
characters, so I reduced it to get the proper length. Also included some
new spellings